factor builder out to top-level class and remove creation work from constructor
