+++ sdk/logs/src/main/java/io/opentelemetry/sdk/logs/SdkLogEmitter.java
@@ -5,13 +5,8 @@
 
 package io.opentelemetry.sdk.logs;
 
-import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
-import io.opentelemetry.sdk.logs.data.Body;
 import io.opentelemetry.sdk.logs.data.LogDataBuilder;
-import io.opentelemetry.sdk.logs.data.Severity;
-import java.time.Instant;
-import java.util.concurrent.TimeUnit;
 
 /** SDK implementation of {@link LogEmitter}. */
 final class SdkLogEmitter implements LogEmitter {
@@ -28,102 +23,16 @@ final class SdkLogEmitter implements LogEmitter {
 
   @Override
   public LogBuilder logBuilder() {
-    return new SdkLogBuilder(logEmitterSharedState, instrumentationLibraryInfo);
+    LogDataBuilder logDataBuilder =
+        LogDataBuilder.create(
+            logEmitterSharedState.getResource(),
+            instrumentationLibraryInfo,
+            logEmitterSharedState.getClock());
+    return new SdkLogBuilder(logEmitterSharedState, logDataBuilder);
   }
 
   // VisibleForTesting
   InstrumentationLibraryInfo getInstrumentationLibraryInfo() {
     return instrumentationLibraryInfo;
-  }
-
-  private static final class SdkLogBuilder implements LogBuilder {
-
-    private final LogDataBuilder logDataBuilder;
-    private final LogEmitterSharedState logEmitterSharedState;
-
-    private SdkLogBuilder(
-        LogEmitterSharedState logEmitterSharedState,
-        InstrumentationLibraryInfo instrumentationLibraryInfo) {
-      this.logEmitterSharedState = logEmitterSharedState;
-      this.logDataBuilder =
-          LogDataBuilder.create(
-              logEmitterSharedState.getResource(),
-              instrumentationLibraryInfo,
-              logEmitterSharedState.getClock());
-    }
-
-    @Override
-    public LogBuilder setEpoch(long timestamp, TimeUnit unit) {
-      logDataBuilder.setEpoch(timestamp, unit);
-      return this;
-    }
-
-    @Override
-    public LogBuilder setEpoch(Instant instant) {
-      logDataBuilder.setEpoch(instant);
-      return this;
-    }
-
-    @Override
-    public LogBuilder setTraceId(String traceId) {
-      logDataBuilder.setTraceId(traceId);
-      return this;
-    }
-
-    @Override
-    public LogBuilder setSpanId(String spanId) {
-      logDataBuilder.setSpanId(spanId);
-      return this;
-    }
-
-    @Override
-    public LogBuilder setFlags(int flags) {
-      logDataBuilder.setFlags(flags);
-      return this;
-    }
-
-    @Override
-    public LogBuilder setSeverity(Severity severity) {
-      logDataBuilder.setSeverity(severity);
-      return this;
-    }
-
-    @Override
-    public LogBuilder setSeverityText(String severityText) {
-      logDataBuilder.setSeverityText(severityText);
-      return this;
-    }
-
-    @Override
-    public LogBuilder setName(String name) {
-      logDataBuilder.setName(name);
-      return this;
-    }
-
-    @Override
-    public LogBuilder setBody(Body body) {
-      logDataBuilder.setBody(body);
-      return this;
-    }
-
-    @Override
-    public LogBuilder setBody(String body) {
-      logDataBuilder.setBody(body);
-      return this;
-    }
-
-    @Override
-    public LogBuilder setAttributes(Attributes attributes) {
-      logDataBuilder.setAttributes(attributes);
-      return this;
-    }
-
-    @Override
-    public void emit() {
-      if (logEmitterSharedState.hasBeenShutdown()) {
-        return;
-      }
-      logEmitterSharedState.getLogProcessor().emit(logDataBuilder.build());
-    }
   }
 }
